Split domain value-object test matrices
